requirements.txt
streamlit
requests
googlesearch-python
Pillow
transformers
groq
torch
